refactor of op ids

+++ fhevm/src/instructions/b_type/bgeu.rs
@@ -5,7 +5,7 @@ use crate::instructions::{decompose, reconstruct};
 pub struct Bgeu();
 
 impl Bgeu {
-    pub fn apply(imm: &[u8; 8], x_rs1: &[u8; 8], x_rs2: &[u8; 8], pc: &[u8; 8]) -> [u8; 8] {
+    pub fn apply(imm: &[u32; 8], x_rs1: &[u32; 8], x_rs2: &[u32; 8], pc: &[u32; 8]) -> [u32; 8] {
         let x_rs1_u32: u32 = reconstruct(x_rs1);
         let x_rs2_u32: u32 = reconstruct(x_rs2);
         if x_rs1_u32 >= x_rs2_u32 {
