+++ fhevm/src/instructions/i_type/slli.rs
@@ -19,7 +19,7 @@ use crate::instructions::{decompose, reconstruct};
 pub struct Slli();
 
 impl Slli {
-    pub fn apply(imm: &[u8; 8], x_rs1: &[u8; 8], _x_rs2: &[u8; 8]) -> [u8; 8] {
+    pub fn apply(imm: &[u32; 8], x_rs1: &[u32; 8], _x_rs2: &[u32; 8]) -> [u32; 8] {
         let imm_u32: u32 = reconstruct(imm);
         let x_rs1_u32: u32 = reconstruct(x_rs1);
         decompose(x_rs1_u32 << std::cmp::min(imm_u32, u32::BITS - 1))
