+++ fhevm/src/instructions/r_type/rem.rs
@@ -3,7 +3,7 @@ use crate::instructions::{decompose, reconstruct};
 pub struct Rem();
 
 impl Rem {
-    pub fn apply(_imm: &[u8; 8], x_rs1: &[u8; 8], x_rs2: &[u8; 8]) -> [u8; 8] {
+    pub fn apply(_imm: &[u32; 8], x_rs1: &[u32; 8], x_rs2: &[u32; 8]) -> [u32; 8] {
         let x_rs1_i32: i32 = reconstruct(x_rs1) as i32;
         let x_rs2_i32: i32 = reconstruct(x_rs2) as i32;
         if x_rs2_i32 == 0 {
