+++ fhevm/src/instructions/r_type/sltu.rs
@@ -2,7 +2,7 @@ use crate::instructions::{decompose, reconstruct};
 pub struct Sltu();
 
 impl Sltu {
-    pub fn apply(_imm: &[u8; 8], x_rs1: &[u8; 8], x_rs2: &[u8; 8]) -> [u8; 8] {
+    pub fn apply(_imm: &[u32; 8], x_rs1: &[u32; 8], x_rs2: &[u32; 8]) -> [u32; 8] {
         if reconstruct(x_rs1) < reconstruct(x_rs2) {
             return decompose(1);
         }
