+++ fhevm/src/instructions/u_type/lui.rs
@@ -21,7 +21,7 @@ use crate::instructions::{decompose, reconstruct};
 pub struct Lui();
 
 impl Lui {
-    pub fn apply(imm: &[u8; 8], _x_rs1: &[u8; 8], _x_rs2: &[u8; 8]) -> [u8; 8] {
+    pub fn apply(imm: &[u32; 8], _x_rs1: &[u32; 8], _x_rs2: &[u32; 8]) -> [u32; 8] {
         decompose(reconstruct(imm))
     }
 }
